requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg==3.2.10
psycopg-binary==3.2.10
PyJWT==2.10.1
sqlparse==0.5.3
asgiref==3.9.1
tzdata==2025.2
dj-database-url==2.1.0
gunicorn==21.2.0
django-cors-headers==4.3.1
